Pass list and include index in results.
Pass the parameters to `search()` as a list rather than a hash reference. This
is more like how the rest of the PGXN::API works (notably construtors).

While at it, add the `index` key to the results hash returned by `search()`.
This is more important now that the index name is no longer used in the URL.